Rename /sparse_dia_qarray.py to /sparsedia_qarray.py
