+++ dynamiqs/qarrays/__init__.py
@@ -1,5 +1,5 @@
 from .dense_qarray import *
 from .qarray import QArray
-from .sparse_dia_qarray import *
+from .sparsedia_qarray import *
 from .type_conversion import *
 from .utils import *
